Assignment 3- uses Zoo.w2
In my Zoo, You can add an animal, put to sleep and wake it, remove all the animals from the zoo( clear ), see all the info about the animals the in the zoo at the current moment, and exit the system. In addition, you can change the screen and choose any other image to put on the screen( if it's in the "photos" package ). For any help in translating this page into English- Please comment here( bellow ).